requirements.txt
Pillow==11.0.0
numpy==2.2.1
opencv-python==4.10.0.84
flask==3.1.0
